backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
psycopg2-binary==2.9.10

